Compute EEG signal parameters

Calculate absolute and power spectrum as well as estimated and lowest frequencies for an EEG signal

fft_eeg(
  eeg_signal,
  sampling_frequency = 125,
  max_frequency = 32,
  num_seconds_window = 5
)

Arguments

eeg_signal

EEG signal expressed in micro-Volts

sampling_frequency

Sampling frequency of the EEG signal. This is typically equal to 125Hz. Default value is 125.

max_frequency

The maximum frequency for which the spectrum is being calculated. Default value is 32.

num_seconds_window

the duration of EEG record used for analysis (in seconds) per window

Value

List containing the elements: - absolute_spectrum: A matrix with every row corresponding to a frequency between 'lowest_frequency' and 'max_frequency'. Every column corresponds to a 5 second interval. Every entry is the absolute value of the Fourier coefficient

- power_spectrum: Same structure as 'absolute_spectrum', but contains the square of the absolute value of the Fourier coefficient

- estimated_frequencies: Vector of estimated frequencies corresponding to each row in the spectrum matrices

- lowest_frequency: The smallest estimable frequency. This is also the first entry of 'estimated_frequencies'